refactor: move oplus exclusive field logic into batt info repo
