+++ app/src/main/java/com/dijia1124/plusplusbattery/service/BatteryMonitorService.kt
@@ -132,7 +132,8 @@ class BatteryMonitorService : Service() {
 
     private suspend fun fetchBatteryStatus(): String = withContext(Dispatchers.IO) {
         val isRoot = prefsRepo.isRootModeFlow.first()
-        val allInfos = batteryRepo.getAvailableBatteryInfo(isRoot)
+        val showOplus = prefsRepo.showOplusFields.first()
+        val allInfos = batteryRepo.getAvailableBatteryInfo(isRoot, showOplus)
         val visibleTypes = prefsRepo.visibleEntriesFlow.first()
         val filtered = if (visibleTypes.isEmpty()) {
             allInfos
